--- environment_win.yml
@@ -1,167 +1,0 @@
-name: CinnFinall
-channels:
-  - defaults
-dependencies:
-  - beautifulsoup4=4.12.3=py312haa95532_0
-  - brotli-python=1.0.9=py312hd77b12b_8
-  - bzip2=1.0.8=h2bbff1b_6
-  - ca-certificates=2024.7.2=haa95532_0
-  - certifi=2024.7.4=py312haa95532_0
-  - charset-normalizer=3.3.2=pyhd3eb1b0_0
-  - expat=2.6.2=hd77b12b_0
-  - idna=3.7=py312haa95532_0
-  - libffi=3.4.4=hd77b12b_1
-  - openssl=3.0.14=h827c3e9_0
-  - pip=24.0=py312haa95532_0
-  - pysocks=1.7.1=py312haa95532_0
-  - python=3.12.4=h14ffc60_1
-  - requests=2.32.3=py312haa95532_0
-  - setuptools=69.5.1=py312haa95532_0
-  - soupsieve=2.5=py312haa95532_0
-  - sqlite=3.45.3=h2bbff1b_0
-  - tk=8.6.14=h0416ee5_0
-  - tzdata=2024a=h04d1e81_0
-  - urllib3=2.0.3=py312haa95532_0
-  - vc=14.2=h2eaa2aa_4
-  - vs2015_runtime=14.29.30133=h43f2093_4
-  - wheel=0.43.0=py312haa95532_0
-  - win_inet_pton=1.1.0=py312haa95532_0
-  - xz=5.4.6=h8cc25b3_1
-  - zlib=1.2.13=h8cc25b3_1
-  - pip:
-      - aiohappyeyeballs==2.3.4
-      - aiohttp==3.10.1
-      - aiosignal==1.3.1
-      - annotated-types==0.7.0
-      - anyio==4.4.0
-      - asgiref==3.8.1
-      - attrs==24.1.0
-      - backoff==2.2.1
-      - bcrypt==4.2.0
-      - build==1.2.1
-      - cachetools==5.4.0
-      - cffi==1.16.0
-      - chroma-hnswlib==0.7.6
-      - click==8.1.7
-      - colorama==0.4.6
-      - coloredlogs==15.0.1
-      - dataclasses-json==0.6.7
-      - deprecated==1.2.14
-      - dirtyjson==1.0.8
-      - distro==1.9.0
-      - environs==9.5.0
-      - fastapi==0.112.0
-      - flatbuffers==24.3.25
-      - frozenlist==1.4.1
-      - google-auth==2.32.0
-      - googleapis-common-protos==1.63.2
-      - greenlet==3.0.3
-      - grpcio==1.63.0
-      - h11==0.14.0
-      - httpcore==1.0.5
-      - httptools==0.6.1
-      - httpx==0.27.0
-      - huggingface-hub==0.24.5
-      - humanfriendly==10.0
-      - importlib-metadata==8.0.0
-      - importlib-resources==6.4.0
-      - joblib==1.4.2
-      - kubernetes==30.1.0
-      - llama-cloud==0.0.12
-      - llama-index==0.10.61
-      - llama-index-agent-openai==0.2.9
-      - llama-index-cli==0.1.13
-      - llama-index-core==0.10.61
-      - llama-index-embeddings-openai==0.1.11
-      - llama-index-indices-managed-llama-cloud==0.2.7
-      - llama-index-legacy==0.9.48
-      - llama-index-llms-openai==0.1.27
-      - llama-index-multi-modal-llms-openai==0.1.8
-      - llama-index-program-openai==0.1.7
-      - llama-index-question-gen-openai==0.1.3
-      - llama-index-readers-file==0.1.32
-      - llama-index-readers-llama-parse==0.1.6
-      - llama-parse==0.4.9
-      - markdown-it-py==3.0.0
-      - marshmallow==3.21.3
-      - mdurl==0.1.2
-      - milvus==2.2.16
-      - milvus-model==0.2.3
-      - mmh3==4.1.0
-      - monotonic==1.6
-      - multidict==6.0.5
-      - mutagen==1.47.0
-      - mypy-extensions==1.0.0
-      - nest-asyncio==1.6.0
-      - nltk==3.8.1
-      - oauthlib==3.2.2
-      - onnxruntime==1.18.1
-      - openai==1.39.0
-      - opentelemetry-api==1.26.0
-      - opentelemetry-exporter-otlp-proto-common==1.26.0
-      - opentelemetry-exporter-otlp-proto-grpc==1.26.0
-      - opentelemetry-instrumentation==0.47b0
-      - opentelemetry-instrumentation-asgi==0.47b0
-      - opentelemetry-instrumentation-fastapi==0.47b0
-      - opentelemetry-proto==1.26.0
-      - opentelemetry-sdk==1.26.0
-      - opentelemetry-semantic-conventions==0.47b0
-      - opentelemetry-util-http==0.47b0
-      - orjson==3.10.6
-      - outcome==1.3.0.post0
-      - overrides==7.7.0
-      - packaging==24.1
-      - posthog==3.5.0
-      - protobuf==4.25.4
-      - pyasn1==0.6.0
-      - pyasn1-modules==0.4.0
-      - pycparser==2.22
-      - pycryptodomex==3.20.0
-      - pydantic==2.8.2
-      - pydantic-core==2.20.1
-      - pydub==0.25.1
-      - pygments==2.18.0
-      - pymilvus==2.4.4
-      - pypdf==4.3.1
-      - pypika==0.48.9
-      - pyproject-hooks==1.1.0
-      - pyreadline3==3.4.1
-      - python-dotenv==1.0.1
-      - pytube==15.0.0
-      - regex==2024.7.24
-      - requests-oauthlib==2.0.0
-      - rich==13.7.1
-      - rsa==4.9
-      - safetensors==0.4.4
-      - scikit-learn==1.5.1
-      - selenium==4.23.1
-      - sentence-transformers==3.0.1
-      - shellingham==1.5.4
-      - sniffio==1.3.1
-      - sortedcontainers==2.4.0
-      - sqlalchemy==2.0.32
-      - starlette==0.37.2
-      - striprtf==0.0.26
-      - tenacity==8.5.0
-      - threadpoolctl==3.5.0
-      - tiktoken==0.7.0
-      - tokenizers==0.19.1
-      - torch==2.4.0
-      - tqdm==4.66.5
-      - transformers==4.43.4
-      - trio==0.26.1
-      - trio-websocket==0.11.1
-      - typer==0.12.3
-      - typing-inspect==0.9.0
-      - ujson==5.10.0
-      - uvicorn==0.30.5
-      - watchfiles==0.22.0
-      - webdriver-manager==4.0.2
-      - websocket-client==1.8.0
-      - websockets==12.0
-      - wrapt==1.16.0
-      - wsproto==1.2.0
-      - yarl==1.9.4
-      - yt-dlp==2024.8.1
-      - zipp==3.19.2
-prefix: C:\Users\Tung\.conda\envs\CinnFinall
