--- requirements.txt
@@ -1,7 +1,0 @@
-requests 
-beautifulsoup4
-yt-dlp
-pymilvus
-xformers
-sentence_transformers
-nltk
