--- text.txt
@@ -1,85 +1,0 @@
-In summer 1997, a full-page ad appeared in The New York Times.
-The message, from the Global Climate Coalition,
-issued a dire economic warning about the US embracing the Kyoto Protocol,
-a treaty to curb greenhouse gas emissions.
-But beneath the veneer of smiling children was something much more insidious:
-a multi-million dollar campaign propped up by questionable data
-and backed by some of the world's most powerful companies.
-The Global Climate Coalition was itself a front for the oil industry—
-established to sow doubt and confusion about climate action.
-But the real story starts decades earlier.
-In the 1970s, oil companies employed some of the world's top atmospheric scientists,
-as they needed to understand weather-related risks to their equipment
-and to assess the environmental impact of new projects.
-By the late 1970s, these scientists, along with their counterparts in academia,
-had concluded that burning fossil fuels created a build up of atmospheric carbon,
-which would impact the climate by trapping heat
-and increasing surface temperatures.
-They warned that an increase of even a few degrees could be catastrophic
-and accurately predicted events such as rapid Arctic warming
-and the melting of Antarctic ice sheets.
-Throughout the 1980s, oil industry reps met repeatedly to discuss these dangers,
-acknowledging the risk that their product posed to the future of humanity.
-However, instead of warning the public,
-or using this knowledge to pivot towards renewable energy sources,
-they doubled down on oil.
-But in the late 1980s, scientists sounded the alarm about climate change,
-raising public awareness, and leading to calls for government action.
-In response, the oil industry launched what would become a decades-long,
-multi-billion-dollar PR campaign
-to discredit the very science they helped pioneer.
-They utilized the same PR firms that had previously helped the tobacco industry
-mislead the public about the harms of smoking.
-Oil companies directly lobbied government officials
-and covertly funded dozens of organizations
-like the Global Climate Coalition,
-whose objective was to obscure the scientific consensus on climate change
-and humanity's role in creating it.
-They attacked credible scientists
-and bankrolled advertisements disguised as op-eds,
-which falsely exaggerated the degree and significance
-of uncertainty in climate models
-and used that uncertainty as an excuse to dismiss the science entirely.
-These "advertorials" grabbed reader's attention with titles like
-"Lies They Tell Our Children," and "Unsettled Science."
-The industry also capitalized on lingering Cold War anxieties
-that equated government regulation with socialism.
-Thus, at the very moment the world was poised to act,
-oil companies shifted the conversation away from the actual science
-and turned it into a debate about protecting freedom.
-By doing so, they took a non-partisan, uncontentious topic
-and transformed it into a hot-button political issue.
-After George W. Bush became president in 2001,
-oil lobbyists successfully pushed his administration to replace officials
-who agreed with mainstream science
-with ones who opposed environmental regulations.
-When Bush pushed the US out of the Kyoto Protocol,
-his administration credited the Global Climate Coalition
-with influencing his decision.
-But the oil industry's PR campaigns didn't end with their Kyoto victory.
-They've continued to shape the climate conversation,
-pushing propaganda and co-opting climate language.
-British Petroleum, for example, popularized the phrase "carbon footprint,"
-an idea which in practice effectively shifts climate responsibility
-from the industry to the consumer.
-To this day, the industry massively overemphasizes their investment
-in green energies, such as biofuels,
-which represent just 1% of their budgets.
-And they employ legions of lobbyists, who attend UN climate meetings
-and work to water down the language of IPCC climate assessment reports.
-In this, they're allied with oil-producing countries,
-which also have a vested interest in continued fossil fuel use.
-While the oil companies now acknowledge that burning fossil fuels
-contributes to climate change,
-they deny having misled the public,
-arguing that their messaging always reflected the scientific consensus.
-But an extensive paper trail shows otherwise.
-While oil companies' profits reach all-time highs,
-climate change costs the public billions of dollars each year.
-Extreme weather events and decreasing air quality kill millions of people annually.
-Meanwhile, the culture of doubt the oil industry created remains widespread,
-polarizing the issue, and delaying meaningful action.
-But it doesn't have to be this way.
-We can still reclaim the conversation and change course,
-embracing renewable energies and sustainable practices
-to protect both our planet and our future.
