--- troll.txt
@@ -1,1 +1,0 @@
-{"thumbnail_height":360,"version":"1.0","url":"https://www.youtube.com/watch?v=rxaLYjC6e9E","thumbnail_width":480,"width":200,"provider_name":"YouTube","type":"video","height":113,"provider_url":"https://www.youtube.com/","html":"<iframe width=\"200\" height=\"113\" src=\"https://www.youtube.com/embed/rxaLYjC6e9E?feature=oembed\" frameborder=\"0\" allow=\"accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share\" referrerpolicy=\"strict-origin-when-cross-origin\" allowfullscreen title=\"Tông môn của ta có triệu điểm mạnh chap 1-22 | Phế Vật Khai Mở Tông Môn Đào Tạo Thần Tiên Thiên Tài\"></iframe>","author_url":"https://www.youtube.com/@kenkunreview","thumbnail_url":"https://i.ytimg.com/vi/rxaLYjC6e9E/hqdefault.jpg","author_name":"KenKun Review","title":"Tông môn của ta có triệu điểm mạnh chap 1-22 | Phế Vật Khai Mở Tông Môn Đào Tạo Thần Tiên Thiên Tài"}
